token in a pile example
In the game of Nim, we start with some number of piles of tokens, and two players alternate in removing one or more tokens from any one pile. The player who takes the last token wins the game.
Write a program that prompts the user to input an integer. This will represent the number of tokens in a pile. The program should print a representation of this pile on the screen. You can do this using string replication.